notifications/notification-constructor-002.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Web Notifications Test: notification - constructor (invalid)</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="author" title="Xin Liu" href="mailto:xinx.liu@intel.com">
<link rel="help" title="5 API" href="http://www.w3.org/TR/notifications/">
<meta name="flags" content="">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
  <div id="log"></div>
  <script>
    test(function() {
      try {
        var notification = new Notification(null);
        assert_unreached("An exception should be thrown");
      } catch (err) {
        assert_true(err != null, "expected exception")
      }
    }, "Test checks that notification constructor arguments can't be null");
  </script>
</body>
